Version 6

+++ app/Http/Controllers/Admin/PostController.php
@@ -53,7 +53,7 @@ class PostController extends Controller
         $post = Post::create($request->all());
 
         if ($request->file('file')) {
-            $url = Storage::put('posts', $request->file('file'));
+            $url = Storage::put('public/posts', $request->file('file'));
 
             $post->image()->create([
                 'url' => $url
@@ -87,7 +87,7 @@ class PostController extends Controller
     public function edit(Post $post)
     {
 
-        $this->autorize('author', '$post');
+        $this->authorize('author', $post);
 
         $categories = Category::pluck('name', 'id');
         $tags = Tag::all();
@@ -105,7 +105,7 @@ class PostController extends Controller
     public function update(PostRequest $request, Post $post)
     {
 
-        $this->autorize('author', '$post');
+        $this->authorize('author', $post);
 
         $post->update($request->all());
 
@@ -141,7 +141,7 @@ class PostController extends Controller
     public function destroy(Post $post)
     {
 
-        $this->autorize('author', '$post');
+        $this->authorize('author', $post);
 
         $post->delete();
